pyproject.toml
[tool.poetry]
name = "fastapi-cdk-demo"
version = "0.1.0"
description = "Demo FastAPI application deployed with CDK"
authors = ["Armand Rego <armand@weirdsheeplabs.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.102.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
